move AssayResultsForSamples components to /entities subpackage
