+++ packages/components/src/entities/SampleActionsButton.tsx
@@ -21,7 +21,7 @@ import { hasAnyPermissions, hasPermissions, User } from '../internal/components/
 import { QueryModel } from '../public/QueryModel/QueryModel';
 import { PicklistCreationMenuItem } from '../internal/components/picklist/PicklistCreationMenuItem';
 import { AddToPicklistMenuItem } from '../internal/components/picklist/AddToPicklistMenuItem';
-import { AssayResultsForSamplesMenuItem } from '../internal/components/assay/AssayResultsForSamplesButton';
+import { AssayResultsForSamplesMenuItem } from './AssayResultsForSamplesButton';
 
 interface Props {
     disabled?: boolean;
